# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant OS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Customer {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email: string | null;
  loyaltyPoints: number;
  totalSpend: number;
  visitCount: number;
  tags: string[];
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CustomerInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  email?: string;
  tags?: string[];
  notes?: string;
}

export interface CustomerUpdate {
  name?: string;
  phone?: string;
  email?: string;
  tags?: string[];
  notes?: string;
}

export interface LoyaltyPointsInput {
  points: number;
  reason?: string;
}

export interface Visit {
  id: number;
  customerId: number;
  /** @nullable */
  orderId?: number | null;
  visitedAt: string;
  amount: number;
  orderType: string;
  /** @nullable */
  notes?: string | null;
}

export interface Table {
  id: number;
  number: number;
  capacity: number;
  status: string;
  section: string;
  /** @nullable */
  notes?: string | null;
}

export interface TableInput {
  number: number;
  capacity: number;
  section: string;
  notes?: string;
}

export interface TableUpdate {
  status?: string;
  capacity?: number;
  section?: string;
  notes?: string;
}

export interface Reservation {
  id: number;
  /** @nullable */
  customerId?: number | null;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  tableId?: number | null;
  partySize: number;
  reservedAt: string;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface ReservationInput {
  customerId?: number;
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  customerPhone: string;
  tableId?: number;
  partySize: number;
  reservedAt: string;
  notes?: string;
}

export interface ReservationUpdate {
  tableId?: number;
  partySize?: number;
  reservedAt?: string;
  status?: string;
  notes?: string;
}

export interface Product {
  id: number;
  name: string;
  price: number;
  category: string;
  /** @nullable */
  description?: string | null;
  available: boolean;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  price: number;
  category: string;
  description?: string;
  available?: boolean;
}

export interface ProductUpdate {
  name?: string;
  price?: number;
  category?: string;
  description?: string;
  available?: boolean;
}

export interface Order {
  id: number;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  tableId?: number | null;
  type: string;
  status: string;
  paymentStatus: string;
  /** @nullable */
  paymentMethod?: string | null;
  totalAmount: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface OrderItem {
  id: number;
  orderId: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  subtotal: number;
  /** @nullable */
  notes?: string | null;
}

export interface OrderDetail {
  id: number;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  tableId?: number | null;
  type: string;
  status: string;
  paymentStatus: string;
  /** @nullable */
  paymentMethod?: string | null;
  totalAmount: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  items: OrderItem[];
}

export type OrderInputItemsItem = {
  productId: number;
  quantity: number;
  notes?: string;
};

export interface OrderInput {
  customerId?: number;
  tableId?: number;
  type: string;
  notes?: string;
  items: OrderInputItemsItem[];
}

export interface OrderUpdate {
  status?: string;
  paymentStatus?: string;
  paymentMethod?: string;
  notes?: string;
}

export interface StaffMember {
  id: number;
  name: string;
  role: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  status: string;
  hireDate: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface StaffInput {
  /** @minLength 1 */
  name: string;
  role: string;
  phone: string;
  email?: string;
  hireDate: string;
  notes?: string;
}

export interface StaffUpdate {
  name?: string;
  role?: string;
  phone?: string;
  email?: string;
  status?: string;
  notes?: string;
}

export interface Shift {
  id: number;
  staffId: number;
  staffName: string;
  date: string;
  startTime: string;
  endTime: string;
  role: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface ShiftInput {
  staffId: number;
  date: string;
  startTime: string;
  endTime: string;
  role: string;
  notes?: string;
}

export interface ShiftUpdate {
  date?: string;
  startTime?: string;
  endTime?: string;
  role?: string;
  notes?: string;
}

export interface Task {
  id: number;
  /** @nullable */
  staffId?: number | null;
  /** @nullable */
  staffName?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  status: string;
  priority: string;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
}

export interface TaskInput {
  staffId?: number;
  /** @minLength 1 */
  title: string;
  description?: string;
  priority: string;
  dueDate?: string;
}

export interface TaskUpdate {
  staffId?: number;
  title?: string;
  description?: string;
  status?: string;
  priority?: string;
  dueDate?: string;
}

export interface DashboardSummary {
  todaySales: number;
  todayOrders: number;
  todayCustomers: number;
  repeatCustomerRate: number;
  pendingOrders: number;
  activeReservations: number;
  weekSales: number;
}

export interface TopProduct {
  productId: number;
  productName: string;
  category: string;
  totalSold: number;
  totalRevenue: number;
}

export interface CustomerFlowPoint {
  hour: string;
  customers: number;
}

export interface ActivityItem {
  id: string;
  type: string;
  description: string;
  /** @nullable */
  amount?: number | null;
  createdAt: string;
}

export type ListCustomersParams = {
  search?: string;
  tag?: string;
};

export type ListReservationsParams = {
  date?: string;
  status?: string;
};

export type ListProductsParams = {
  category?: string;
};

export type ListOrdersParams = {
  status?: string;
  type?: string;
  date?: string;
};

export type ListStaffParams = {
  role?: string;
};

export type ListShiftsParams = {
  date?: string;
  staffId?: number;
};

export type ListTasksParams = {
  staffId?: number;
  status?: string;
};