# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Restaurant OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: tag
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Customer" }
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /customers/{id}/visits:
    get:
      operationId: listCustomerVisits
      tags: [customers]
      summary: Get visit history for a customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Visit list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Visit" }
  /customers/{id}/add-points:
    post:
      operationId: addLoyaltyPoints
      tags: [customers]
      summary: Add loyalty points to a customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoyaltyPointsInput" }
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }

  # ─── TABLES ──────────────────────────────────────────────────
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List menu products
      parameters:
        - name: category
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── ORDERS ───────────────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema: { type: string }
        - name: date
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order with items
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order with items
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderDetail" }
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status or payment
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }

  # ─── STAFF ────────────────────────────────────────────────────
...
components:
  schemas:
...
    Customer:
      type: object
      required: [id, name, phone, email, loyaltyPoints, totalSpend, visitCount, tags, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        loyaltyPoints: { type: integer }
        totalSpend: { type: number }
        visitCount: { type: integer }
        tags: { type: array, items: { type: string } }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    CustomerInput:
      type: object
      required: [name, phone]
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string, minLength: 1 }
        email: { type: string }
        tags: { type: array, items: { type: string } }
        notes: { type: string }
    CustomerUpdate:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        email: { type: string }
        tags: { type: array, items: { type: string } }
        notes: { type: string }
    LoyaltyPointsInput:
      type: object
      required: [points]
      properties:
        points: { type: integer }
        reason: { type: string }
    Visit:
      type: object
      required: [id, customerId, visitedAt, amount, orderType]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        orderId: { type: ["integer", "null"] }
        visitedAt: { type: string }
        amount: { type: number }
        orderType: { type: string }
        notes: { type: ["string", "null"] }

    # ── Table ──
...
    Product:
      type: object
      required: [id, name, price, category, available, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        price: { type: number }
        category: { type: string }
        description: { type: ["string", "null"] }
        available: { type: boolean }
        createdAt: { type: string }
    ProductInput:
      type: object
      required: [name, price, category]
      properties:
        name: { type: string, minLength: 1 }
        price: { type: number }
        category: { type: string }
        description: { type: string }
        available: { type: boolean }
    ProductUpdate:
      type: object
      properties:
        name: { type: string }
        price: { type: number }
        category: { type: string }
        description: { type: string }
        available: { type: boolean }

    # ── Order ──
    Order:
      type: object
      required: [id, type, status, paymentStatus, totalAmount, createdAt]
      properties:
        id: { type: integer }
        customerId: { type: ["integer", "null"] }
        tableId: { type: ["integer", "null"] }
        type: { type: string }
        status: { type: string }
        paymentStatus: { type: string }
        paymentMethod: { type: ["string", "null"] }
        totalAmount: { type: number }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    OrderItem:
      type: object
      required: [id, orderId, productId, productName, quantity, unitPrice, subtotal]
      properties:
        id: { type: integer }
        orderId: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        subtotal: { type: number }
        notes: { type: ["string", "null"] }
    OrderDetail:
      type: object
      required: [id, type, status, paymentStatus, totalAmount, items, createdAt]
      properties:
        id: { type: integer }
        customerId: { type: ["integer", "null"] }
        tableId: { type: ["integer", "null"] }
        type: { type: string }
        status: { type: string }
        paymentStatus: { type: string }
        paymentMethod: { type: ["string", "null"] }
        totalAmount: { type: number }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
    OrderInput:
      type: object
      required: [type, items]
      properties:
        customerId: { type: integer }
        tableId: { type: integer }
        type: { type: string }
        notes: { type: string }
        items:
          type: array
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId: { type: integer }
              quantity: { type: integer }
              notes: { type: string }
    OrderUpdate:
      type: object
      properties:
        status: { type: string }
        paymentStatus: { type: string }
        paymentMethod: { type: string }
        notes: { type: string }

    # ── Staff ──
...