# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant OS API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerUpdate {
  name?: string;
  phone?: string;
  email?: string;
  tags?: string[];
  notes?: string;
}