# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant OS API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  tableId?: number | null;
  type: string;
  status: string;
  paymentStatus: string;
  /** @nullable */
  paymentMethod?: string | null;
  totalAmount: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}