# lib/api-zod/src/generated/types/visit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant OS API
 * OpenAPI spec version: 0.1.0
 */

export interface Visit {
  id: number;
  customerId: number;
  /** @nullable */
  orderId?: number | null;
  visitedAt: string;
  amount: number;
  orderType: string;
  /** @nullable */
  notes?: string | null;
}